Extract project selection logic.
